Replaced Audiofiles, reconnected Scripts
