--- Assets/_Project/Audio/General/Player/Mechanics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 86bbdb345d8bcde4880a201e76664b40
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
