--- Assets/_Project/Core/Scripts/_GameScripts/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5365a93fe17a4374484e432780ec63a6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
