--- Assets/_Project/Core/Scripts/_GameScripts/Framework.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d3e476d758a827f4fbe1680d2508f5df
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
