--- Assets/_Project/Core/Scripts/_GameScripts/Gameplay/Audio/Alexej/AudioRadiation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 18ce39b976096254687dbe93d48a257f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
