--- Assets/_Project/Core/Scripts/_GameScripts/Gameplay/Audio/Alexej/MuteMusic.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c87dfc36cb7cd3f459b24ca106f226ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
