--- Assets/_Project/Core/Scripts/_GameScripts/Gameplay/Audio/Alexej/SnapshotChange.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c509807e45043e74daa12cfe997655c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
