--- Assets/_Project/Core/Scripts/_GameScripts/Graphics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 626a709e8377d7d48957adffe8b995a8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
