+++ Assets/_Prototype/Loudness x Silence/Audio/Mixers/Radiation+Hub.mixer
@@ -43,7 +43,7 @@ AudioMixerSnapshotController:
   m_AudioMixer: {fileID: 24100000}
   m_SnapshotID: fb84354d4fc8c4b60989d4055ad5874e
   m_FloatValues:
-    d5c71224884344a5f92b6716c94c5f42: -8.010123
+    d5c71224884344a5f92b6716c94c5f42: -80
     1279af565bb594c40b129f9782d0082f: -80
     e1d53786bbff14566a3927403b694e2e: 0
     e60d6059a169540d6a57cfa8dea44d36: -80
@@ -81,8 +81,8 @@ AudioMixerSnapshotController:
   m_FloatValues:
     d5c71224884344a5f92b6716c94c5f42: 0
     1279af565bb594c40b129f9782d0082f: -80
-    e1d53786bbff14566a3927403b694e2e: -8.481216
-    e60d6059a169540d6a57cfa8dea44d36: -11.535833
+    e1d53786bbff14566a3927403b694e2e: 0
+    e60d6059a169540d6a57cfa8dea44d36: -80
   m_TransitionOverrides:
     d5c71224884344a5f92b6716c94c5f42: 1
     1279af565bb594c40b129f9782d0082f: 1
@@ -187,7 +187,7 @@ AudioMixerSnapshotController:
   m_SnapshotID: 85d74b4e2bd7b49ac83f4dcd560a8aed
   m_FloatValues:
     d5c71224884344a5f92b6716c94c5f42: -80
-    1279af565bb594c40b129f9782d0082f: -80
+    1279af565bb594c40b129f9782d0082f: 0.23288462
     e1d53786bbff14566a3927403b694e2e: -80
     e60d6059a169540d6a57cfa8dea44d36: -80
   m_TransitionOverrides:
@@ -244,9 +244,9 @@ AudioMixerSnapshotController:
   m_AudioMixer: {fileID: 24100000}
   m_SnapshotID: 1e888f7ec55f748b6b3b913a878a37e2
   m_FloatValues:
-    d5c71224884344a5f92b6716c94c5f42: -8.0101185
+    d5c71224884344a5f92b6716c94c5f42: 0
     1279af565bb594c40b129f9782d0082f: -80
-    e1d53786bbff14566a3927403b694e2e: -80
+    e1d53786bbff14566a3927403b694e2e: 0
     e60d6059a169540d6a57cfa8dea44d36: 0
   m_TransitionOverrides: {}
 --- !u!243 &8471017030050929498
