+++ Assets/_Prototype/Loudness x Silence/Audio/Mixers/dearVRMixer.mixer
@@ -169,7 +169,9 @@ AudioMixerSnapshotController:
     f066148d2d5544869aa6ad76a1417397: 0
     e31c1dbe6e16a43efb84bd13b197f84f: 0
     676ecbff83aaa442995d7e2eed577e7b: 79.7
-  m_TransitionOverrides: {}
+  m_TransitionOverrides:
+    1bc68491c7dd44bd8b63b0a0f22e2ade: 1
+    9a94e04bbf6164ec19b6760c3b959942: 1
 --- !u!244 &206494336722648633
 AudioMixerEffectController:
   m_ObjectHideFlags: 3
