--- Assets/_Prototype/Loudness x Silence/Audio/NoiseWolrd/NoiseWorld_2.wav.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: 2bb2cfb7da1c34515beaaa9ad9593c17
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
