--- Assets/_Prototype/Loudness x Silence/Audio/NoiseWolrd/NoiseWorld_3.wav.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: af653d98ad5704129b64e22e8e767669
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
