+++ Assets/_Prototype/Loudness x Silence/Audio/NoiseWolrd/mx_uni_radiation_layer_a.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 0d7c027f0158a4136bb578202cc33500
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
