+++ Assets/_Prototype/Loudness x Silence/Audio/NoiseWolrd/mx_uni_radiation_layer_c.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: a90684a392b844506a172478f5305338
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
