+++ Assets/_Prototype/Loudness x Silence/Scene/Loudness x Silence.unity
@@ -994,7 +994,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 739300859}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 61f430e58da8545e48af2b810d69cb17, type: 3}
+  m_Script: {fileID: 11500000, guid: 18ce39b976096254687dbe93d48a257f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _zero: {fileID: 24500006, guid: bcbc931174e5a432db16e96090ad5896, type: 2}
@@ -1003,7 +1003,7 @@ MonoBehaviour:
   _phase2: {fileID: -1853917020019906566, guid: bcbc931174e5a432db16e96090ad5896,
     type: 2}
   _phase3: {fileID: 7539695448031531986, guid: bcbc931174e5a432db16e96090ad5896, type: 2}
-  _TransitionIn: 3
+  _TransitionIn: 4
   _Loudness: 0
 --- !u!4 &739300861
 Transform:
@@ -1029,15 +1029,15 @@ MonoBehaviour:
   m_GameObject: {fileID: 739300859}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fa999352d3f6d454fa1c0cfaced28477, type: 3}
+  m_Script: {fileID: 11500000, guid: c509807e45043e74daa12cfe997655c1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _3D: {fileID: 24500006, guid: fda9214ce9b1346cdb2844990b7c9aa1, type: 2}
   _NO3D: {fileID: -1668553965950473002, guid: fda9214ce9b1346cdb2844990b7c9aa1, type: 2}
-  _hubMixer: {fileID: 24100000, guid: bcbc931174e5a432db16e96090ad5896, type: 2}
+  _hubMixer: {fileID: 24100000, guid: fda9214ce9b1346cdb2844990b7c9aa1, type: 2}
   _TransitionIn: 1
   _hubVol: -80
-  _mixTimer: 1
+  _mixTimer: 0
 --- !u!114 &739300863
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1047,7 +1047,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 739300859}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c12f0550a688e46458a61d6519a0c0b1, type: 3}
+  m_Script: {fileID: 11500000, guid: c87dfc36cb7cd3f459b24ca106f226ec, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _MusicMixer: {fileID: 24100000, guid: fda9214ce9b1346cdb2844990b7c9aa1, type: 2}
@@ -1750,7 +1750,7 @@ AudioSource:
   serializedVersion: 4
   OutputAudioMixerGroup: {fileID: -7719575150818130832, guid: bcbc931174e5a432db16e96090ad5896,
     type: 2}
-  m_audioClip: {fileID: 8300000, guid: 45082bbf731d142bf939b9960fd6313b, type: 3}
+  m_audioClip: {fileID: 8300000, guid: 764470171553643c2b0db4e78289f5ab, type: 3}
   m_PlayOnAwake: 1
   m_Volume: 0.2
   m_Pitch: 1
@@ -1926,7 +1926,7 @@ AudioSource:
   serializedVersion: 4
   OutputAudioMixerGroup: {fileID: 8471017030050929498, guid: bcbc931174e5a432db16e96090ad5896,
     type: 2}
-  m_audioClip: {fileID: 8300000, guid: af653d98ad5704129b64e22e8e767669, type: 3}
+  m_audioClip: {fileID: 8300000, guid: a90684a392b844506a172478f5305338, type: 3}
   m_PlayOnAwake: 1
   m_Volume: 0.2
   m_Pitch: 1
@@ -2054,7 +2054,7 @@ AudioSource:
   serializedVersion: 4
   OutputAudioMixerGroup: {fileID: -1715796203902366957, guid: bcbc931174e5a432db16e96090ad5896,
     type: 2}
-  m_audioClip: {fileID: 8300000, guid: c17d4f524435c42378d7a91786edb9e0, type: 3}
+  m_audioClip: {fileID: 8300000, guid: 0d7c027f0158a4136bb578202cc33500, type: 3}
   m_PlayOnAwake: 1
   m_Volume: 0.2
   m_Pitch: 1
