+++ Assets/_Prototype/Prototyping.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9473a1588a64a46e4b590f17d47ed552
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
