--- Assets/__NoVersionControl.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 08bf715256bcfc84ebeb0d7d51958b72
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
